requirements.txt
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.14.2
numpy==1.23.1
pandas==1.4.3

